+++ firmware/CHANGELOG.md
@@ -39,6 +39,7 @@ Release template (copy/paste this for new release):
  - VE/Ign/Boost blend tables can select a Y axis override for even more flexibility
  - Disable radiator fans after certain vehicle speed #5875
  - LIS2DW12TR on-board Accelerometer support #4870
+ - Wheel slip ratio using aux speed inputs #5897
 
 ### Removed
  - Narrow to Wideband approximation